pyproject.toml
[project]
name = "monomer-hackathon-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.0",
    "bio>=1.8.1",
    "fastapi[standard]>=0.120.0",
    "futurehouse-client>=0.6.7",
    "ipykernel>=7.0.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-openai>=0.3.32",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "scikit-optimize>=0.10.2",
    "scipy>=1.16.2",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.50.0",
    "uvicorn>=0.38.0",
]
